First pass at CLI DynamoDB distributed lock
Initial commit that wraps ddbsync with a cli arg parser to create a
distributed locking CLI tool using DynamoDB. Flags are passed on the CLI
or as ENV vars with sane defaults. Initially this was supposed to be a
blocking locking mechanism that passed arguments straight through to the
lock and deferred unlocking until the command returned. Variadic
functions or arg parsers are more complex that I’d have liked for a
first pass. What if the arg command has some bad characters? Spans
multiple lines? What should the lock do with exit codes of the command
and arguments passed to it? How do we pass in arguments we do want
i.e. lock name? All while allowing the operator to know what
commands are accepted etc.

For these reasons, I have instead opted for creating two possible
Dynolocker commands: lock/unlock. Where we need to call this lock, we
can lock and then unlock (or wait for the lock to expire).

By default the lock has a TTL of 60 seconds. Another lock will retry
indefinitely if unable to acquire a lock, retrying every 10 seconds
until a) the lock expires b) is manually unlocked.